+++ example/thing.model.dart
@@ -1,0 +1,4 @@
+class Thing {
+  const Thing({required this.thing});
+  final String thing;
+}