+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-08-25
+
+### Added
+- New extension method `indexOfOrNull(value)`  
+  Returns the index of the first occurrence of a given value in the iterable, or `null` if the value is not found.
+
+---
+
 ## [1.0.0] - 2025-08-23
 
 ### Added